requirements.txt
tensorflow==2.12
numpy==1.23.5
streamlit==1.29.0
pillow==9.5.0